Move helpers to utils directory
